requirements.txt
google-generativeai
pillow
opencv-python
numpy
matplotlib
requests
